download links and footer

+++ app/assets/javascripts/controllers/documents_controller.js.coffee
@@ -22,11 +22,13 @@ Zapply.DocumentsController = Ember.ArrayController.extend
     # Zapply.Document.find()
 
   destroyDocument: (document) ->
-    document.deleteRecord()
-    @get('store').commit()
+    confirmation = confirm("Are you sure you want to delete " + document.get('material') + "?")
+    if confirmation
+      document.deleteRecord()
+      @get('store').commit()
 
   submitApplication: ->
-    session = @get('session').get('content')
-    session.set("submitted", true)
-    @get('store').commit()
-    console.log "submitted: ", session.get('submitted')
+    if confirm("Once you submit your application, it will be transmitted to WolfBrown and you will no longer be able to edit it. Are you sure you want to proceed?")
+      session = @get('session').get('content')
+      session.set("submitted", true)
+      @get('store').commit()
